Added functionality to Exx and Exy calculations
Now, the terms Exx and Exy takes into account the possibility that the empiric communities results in just one cluster. A random example given this situation results in a overlaping Rcom value of 1, meaning that the formation of the obseved cluster is random without regards of how the edges are permuted between the nodes.